second part

+++ app/Http/Controllers/AccountsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Account;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use \Illuminate\Auth\Middleware\Authorize;
 
 class AccountsController extends Controller
 {
@@ -28,9 +29,13 @@ class AccountsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create(Account $account)
     {
-        return view('accounts.create');
+       
+        
+            $this->authorize('create-accounts', $account);
+            return view('accounts.create');
+        
     }
 
     /**
@@ -62,6 +67,7 @@ class AccountsController extends Controller
         //$id = accounts::get
        // $accounts = DB::table('accounts')->find(1);
         //$accounts = Account::paginate(10);
+        $this->authorize('show-accounts', $account);
         return view('accounts.show',compact('account'));
 
 
@@ -75,6 +81,8 @@ class AccountsController extends Controller
     public function edit(Account $account)
     {
 
+      $this->authorize('update-accounts', $account);
+        
         return view('accounts.edit',compact('account'));
     }
 
@@ -90,6 +98,7 @@ class AccountsController extends Controller
 
     public function update(Request $request, Account $account)
     {
+        
         $request->validate([
             'name' => 'required',
 
@@ -109,6 +118,7 @@ class AccountsController extends Controller
      */
     public function destroy(Account $account)
     {
+        $this->authorize('delete-accounts', $account);
        $account->delete();
